// src/types/category.enum.ts
export enum CATEGORY {
  EDUCATION = 'Education',
  HEALTH = 'Health',
  ANIMAL_WELFARE = 'Animal Welfare',
  HUMAN_SERVICES = 'Human Services',
  ARTS_CULTURE = 'Arts & Culture',
  ENVIRONMENT = 'Environment',
}

export enum CATEGORY_SLUGS {
  EDUCATION = 'education',
  HEALTH = 'health',
  ANIMAL = 'animal',
  HUMAN = 'human',
  ARTS = 'arts',
  ENVIRONMENT = 'environment',
}